chore: add helper functions to dedicated file
